requirements.txt
docutils==0.12
lockfile==0.12.2
mailchimp3==1.0.26
python-daemon==2.1.1
pyzmq==15.4.0
requests==2.11.1
wheel==0.24.0
